microservices/pdf-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.15.0
PyMuPDF==1.23.26
pika==1.3.2
bson==0.5.10
python-dotenv==1.0.0
colorama==0.4.6